+++ .ladle/components.tsx
@@ -1,8 +1,10 @@
-import React from 'react'
+// organize-imports-ignore
 import type { GlobalProvider } from '@ladle/react'
+import React from 'react'
 
-import ThemeProvider from '../src/ThemeProvider'
+import 'tailwindcss/src/css/preflight.css'
 import 'tailwindcss/tailwind.css'
+import ThemeProvider from '../src/ThemeProvider'
 
 export const Provider: GlobalProvider = ({ children, globalState }) => (
   <>