+++ src/Prose.stories.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import Prose from './Prose'
 
 export const Default = () => {