+++ src/constants.js
@@ -1,3 +1,50 @@
+const TailwindAnimations = {
+  keyframes: {
+    'enter-from-right': {
+      '0%': { transform: 'translateX(200px)', opacity: 0 },
+      '100%': { transform: 'translateX(0)', opacity: 1 },
+    },
+    'enter-from-left': {
+      '0%': { transform: 'translateX(-200px)', opacity: 0 },
+      '100%': { transform: 'translateX(0)', opacity: 1 },
+    },
+    'exit-to-right': {
+      '0%': { transform: 'translateX(0)', opacity: 1 },
+      '100%': { transform: 'translateX(200px)', opacity: 0 },
+    },
+    'exit-to-left': {
+      '0%': { transform: 'translateX(0)', opacity: 1 },
+      '100%': { transform: 'translateX(-200px)', opacity: 0 },
+    },
+    'scale-in-content': {
+      '0%': { transform: 'rotateX(-30deg) scale(0.9)', opacity: 0 },
+      '100%': { transform: 'rotateX(0deg) scale(1)', opacity: 1 },
+    },
+    'scale-out-content': {
+      '0%': { transform: 'rotateX(0deg) scale(1)', opacity: 1 },
+      '100%': { transform: 'rotateX(-10deg) scale(0.95)', opacity: 0 },
+    },
+    'fade-in': {
+      '0%': { opacity: 0 },
+      '100%': { opacity: 1 },
+    },
+    'fade-out': {
+      '0%': { opacity: 1 },
+      '100%': { opacity: 0 },
+    },
+  },
+  animation: {
+    'enter-from-right': 'enter-from-right 0.25s ease',
+    'enter-from-left': 'enter-from-left 0.25s ease',
+    'exit-to-right': 'exit-to-right 0.25s ease',
+    'exit-to-left': 'exit-to-left 0.25s ease',
+    'scale-in-content': 'scale-in-content 0.2s ease',
+    'scale-out-content': 'scale-out-content 0.2s ease',
+    'fade-in': 'fade-in 0.2s ease',
+    'fade-out': 'fade-out 0.2s ease',
+  },
+}
+
 const TailwindColorsUZH = {
   'uzh-blue-100': '#0028a5',
   'uzh-blue-80': '#3353b7',
@@ -60,7 +107,8 @@ const PrettierConfig = {
   singleQuote: true,
   trailingComma: 'es5',
   plugins: [
-    require('prettier-plugin-tailwindcss'),
+    // FIXME: currently cannot use tailwindcss together with organize imports (https://github.com/tailwindlabs/prettier-plugin-tailwindcss/issues/31)
+    // require('prettier-plugin-tailwindcss'),
     require('prettier-plugin-organize-imports'),
   ],
 }
@@ -125,6 +173,7 @@ const VersionConfigTypes = [
 ]
 
 module.exports = {
+  TailwindAnimations,
   TailwindColorsUZH,
   TailwindFonts,
   TailwindBasePlugins,