+++ src/index.ts
@@ -1,7 +1,13 @@
 import './tailwind.css'
 
 export * from './Button'
+export * from './Countdown'
 export * from './Header'
 export * from './Modal'
+export * from './Navigation'
+export * from './NotificationBadgeWrapper'
+export * from './Progress'
 export * from './Prose'
+export * from './Slider'
+export * from './Switch'
 export * from './ThemeProvider'