+++ CHANGELOG.md
@@ -2,6 +2,158 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.37](https://github.com/uzh-bf/design-system/compare/v0.0.34...v0.0.37) (2022-09-12)
+
+
+### Features
+
+* **NotificationBadgeWrapper:** add notification badge wrapper component ([#10](https://github.com/uzh-bf/design-system/issues/10)) ([6b8bff4](https://github.com/uzh-bf/design-system/commit/6b8bff4ed86a3224954d72090f87d6be110dfc85))
+
+
+### Other
+
+* **release:** 0.0.35 ([18b98c9](https://github.com/uzh-bf/design-system/commit/18b98c935d7c7c3acfa233a892cffdd3d6a3b888))
+* **release:** 0.0.36 ([1f8ee34](https://github.com/uzh-bf/design-system/commit/1f8ee34805e1b09b43349a338f325bfececa6dc8))
+
+
+### Enhancements
+
+* add slider component ([#11](https://github.com/uzh-bf/design-system/issues/11)) ([cdda43b](https://github.com/uzh-bf/design-system/commit/cdda43b3d2ce3a3b10b0f075b02980c878632c98))
+* **Progress:** make max number optional ([9ae2fa6](https://github.com/uzh-bf/design-system/commit/9ae2fa6616572851032db7c310cd365c663d3d8d))
+
+### [0.0.36](https://github.com/uzh-bf/design-system/compare/v0.0.35...v0.0.36) (2022-09-11)
+
+
+### Enhancements
+
+* add slider component ([#11](https://github.com/uzh-bf/design-system/issues/11)) ([cdda43b](https://github.com/uzh-bf/design-system/commit/cdda43b3d2ce3a3b10b0f075b02980c878632c98))
+
+### [0.0.35](https://github.com/uzh-bf/design-system/compare/v0.0.34...v0.0.35) (2022-09-08)
+
+
+### Features
+
+* **NotificationBadgeWrapper:** add notification badge wrapper component ([#10](https://github.com/uzh-bf/design-system/issues/10)) ([6b8bff4](https://github.com/uzh-bf/design-system/commit/6b8bff4ed86a3224954d72090f87d6be110dfc85))
+
+### [0.0.34](https://github.com/uzh-bf/design-system/compare/v0.0.33...v0.0.34) (2022-09-06)
+
+### [0.0.33](https://github.com/uzh-bf/design-system/compare/v0.0.32...v0.0.33) (2022-09-06)
+
+### [0.0.32](https://github.com/uzh-bf/design-system/compare/v0.0.31...v0.0.32) (2022-09-06)
+
+
+### Other
+
+* **Markdown:** remove markdown component and stories due to dependency conflicts on package import ([#9](https://github.com/uzh-bf/design-system/issues/9)) ([1c02d4b](https://github.com/uzh-bf/design-system/commit/1c02d4bf3c5f516726a412ecd26e59080957e192))
+
+### [0.0.31](https://github.com/uzh-bf/design-system/compare/v0.0.30...v0.0.31) (2022-09-05)
+
+
+### Features
+
+* **Markdown:** add markdown component ([#8](https://github.com/uzh-bf/design-system/issues/8)) ([d8fd94e](https://github.com/uzh-bf/design-system/commit/d8fd94e34b16067d9fbce93abc1eeac8f1e056de))
+
+### [0.0.30](https://github.com/uzh-bf/design-system/compare/v0.0.29...v0.0.30) (2022-09-02)
+
+
+### Bug Fixes
+
+* **Countdown:** add missing exports of countdown component ([ee3c9a4](https://github.com/uzh-bf/design-system/commit/ee3c9a49b65a97206cc0707bd78547a4d94fb3cb))
+
+### [0.0.29](https://github.com/uzh-bf/design-system/compare/v0.0.28...v0.0.29) (2022-09-02)
+
+
+### Other
+
+* added missing export for Countdown component ([dc23426](https://github.com/uzh-bf/design-system/commit/dc23426d32c94c944b1e0c976cc3123156daee38))
+
+### [0.0.28](https://github.com/uzh-bf/design-system/compare/v0.0.27...v0.0.28) (2022-09-02)
+
+
+### Features
+
+* **Countdown:** add countdown component ([58c7a75](https://github.com/uzh-bf/design-system/commit/58c7a75b6d7560dd99f145a6a4f5febd30d1a0fa))
+
+
+### Bug Fixes
+
+* fixed layout issues for progress bar and added corresponding story ([2a13641](https://github.com/uzh-bf/design-system/commit/2a13641a536d853092bbdbbe0133fa2697820fce))
+
+
+### Enhancements
+
+* implement Countdown and corresponding stories ([da963b0](https://github.com/uzh-bf/design-system/commit/da963b0c83d4b6275ac3c062b5ba608e98819944))
+
+### [0.0.27](https://github.com/uzh-bf/design-system/compare/v0.0.25...v0.0.27) (2022-09-02)
+
+
+### Features
+
+* **Progress:** add progress bar component and upgrade packages ([#4](https://github.com/uzh-bf/design-system/issues/4)) ([52db72d](https://github.com/uzh-bf/design-system/commit/52db72dc5091b4a5ccccdcf07f5e5c28b6c55b7b))
+* **Switch:** add switch component ([#5](https://github.com/uzh-bf/design-system/issues/5)) ([5921151](https://github.com/uzh-bf/design-system/commit/5921151464a21eb892ecbbb1b533456394d92e4a))
+
+
+### Bug Fixes
+
+* fix linter warnings for navbar ([1630d87](https://github.com/uzh-bf/design-system/commit/1630d87db77b26325c1646845c592a31ca6a7eb0))
+
+
+### Other
+
+* allow for color and height changes ([c2af5a3](https://github.com/uzh-bf/design-system/commit/c2af5a3e4461d857e8aaa68ce3088d5ae139bf2a))
+* fix naming scheme ([48e0a3a](https://github.com/uzh-bf/design-system/commit/48e0a3aac5855cdafb158f5affaa6560b0207855))
+* **release:** 0.0.26 ([f7d8cf8](https://github.com/uzh-bf/design-system/commit/f7d8cf88b12569320a462fea77e2afa3c09b139d))
+
+### [0.0.26](https://github.com/uzh-bf/design-system/compare/v0.0.25...v0.0.26) (2022-08-22)
+
+
+### Features
+
+* **Progress:** add progress bar component and upgrade packages ([#4](https://github.com/uzh-bf/design-system/issues/4)) ([52db72d](https://github.com/uzh-bf/design-system/commit/52db72dc5091b4a5ccccdcf07f5e5c28b6c55b7b))
+* **Switch:** add switch component ([#5](https://github.com/uzh-bf/design-system/issues/5)) ([5921151](https://github.com/uzh-bf/design-system/commit/5921151464a21eb892ecbbb1b533456394d92e4a))
+
+### [0.0.25](https://github.com/uzh-bf/design-system/compare/v0.0.24...v0.0.25) (2022-08-08)
+
+### [0.0.24](https://github.com/uzh-bf/design-system/compare/v0.0.23...v0.0.24) (2022-08-08)
+
+
+### Bug Fixes
+
+* **Navigation:** add export for component ([6ec1c9d](https://github.com/uzh-bf/design-system/commit/6ec1c9d5f198ab2231d28915b5c76fd94eda3b88))
+
+### [0.0.23](https://github.com/uzh-bf/design-system/compare/v0.0.22...v0.0.23) (2022-08-08)
+
+
+### Features
+
+* **Navigation:** add navigation based on Radix NavigationMenu [#2](https://github.com/uzh-bf/design-system/issues/2)  ([69ce9c7](https://github.com/uzh-bf/design-system/commit/69ce9c7203ab8e75af2fc69c6e16d4fa742de131))
+
+
+### Refactors
+
+* extract animation config to constants ([a25f1ba](https://github.com/uzh-bf/design-system/commit/a25f1bad268fc334cc63e9fef7879818fe718877))
+
+### [0.0.22](https://github.com/uzh-bf/design-system/compare/v0.0.21...v0.0.22) (2022-08-08)
+
+
+### Bug Fixes
+
+* disable prettier-plugin-tailwindcss for now (breaks with organize imports) ([4cfff3a](https://github.com/uzh-bf/design-system/commit/4cfff3a47186264ad4268282df7d78939e5aaf74))
+
+
+### Other
+
+* remove all peer deps to allow usage in the backend ([1f3b87d](https://github.com/uzh-bf/design-system/commit/1f3b87dd7bc98f4e3ed9b0639797e02439174d64))
+
+### [0.0.21](https://github.com/uzh-bf/design-system/compare/v0.0.20...v0.0.21) (2022-08-08)
+
+### [0.0.20](https://github.com/uzh-bf/design-system/compare/v0.0.19...v0.0.20) (2022-08-08)
+
+
+### Other
+
+* disable preflight globally (import on case by case basis), remove peer deps for code style tooling ([cf13678](https://github.com/uzh-bf/design-system/commit/cf13678c91e2e8e94d5321c8ad7ffd9bf63669f6))
+
 ### [0.0.19](https://github.com/uzh-bf/design-system/compare/v0.0.18...v0.0.19) (2022-07-23)